+++ out/editors/baseEditor.js
@@ -96,6 +96,18 @@ class BaseCustomTextEditorProvider {
         }
         return changeDocumentSubscription;
     }
+    /**
+     * Open a file relative to the current document.
+     * @param currentDoc The current document
+     * @param relativePath The relative path to the file to open
+     */
+    openFile(currentDoc, relativePath) {
+        if (!relativePath) {
+            return;
+        }
+        const targetUri = vscode.Uri.joinPath(currentDoc.uri, '..', relativePath);
+        vscode.commands.executeCommand('vscode.open', targetUri);
+    }
 }
 exports.BaseCustomTextEditorProvider = BaseCustomTextEditorProvider;
 //# sourceMappingURL=baseEditor.js.map